requirements.txt
streamlit
yfinance==0.2.28
plotly
kaleido
ollama
python-dotenv